Enable Strict Transport Security on hydra.nixos.org and nixos.org
HSTS tells the browser that you should always use HTTPS when
talking to the domain, for the next max-age. This means it will
(in-browser) redirect from http to https, without hitting a
the webserver.

 - max-age is set to 1 day in case of problems, in the future I'd
   like to extend this.
 - we can't enable includeSubdomains because the following
   certificates don't use TLS:

   - www.nixos.org (redirects to nixos.org)
   - releases.nixos.org (deprecated? traffic?
   - planet.nixos.org

  If we can fix these, we can add includeSubdomains too.

Obviously I haven't tested this, as I don't have a local copy of
nixos.org running :)