delft/eris: update disk monitoring alerts
Quite recently the `/scratch` partition on bastion ran out of free disk
space.

This is an update to the alerting rule to generalize and extend the
existing disk fullness alerts.

In addition to monitoring `/scratch` we are now sharing the filesystem
selector between these rules and in addition monitor relative fullness
on them.