Drop virtualized macs
Using qemu for the macs seemed to work well for a while, but ultimately
became really hard to manage. Especially with hardwarde and software
upgrades which became more and more hostile to the idea.

Once we get the hardware, owned macs back online we can reinstall
macOS and enroll it in MDM instead.

This commit also includes a nix-darwin.nix file, which is
applied with MDM and
https://github.com/DeterminateSystems/macos-ephemeral.