Introduce reproducible.nixos.org
It would be good to have a place to collect information about
Reproducible Builds on Nix. Currently there is https://r13y.com,
but it is rather outdated and it would be good to gradually move
this to more shared infrastructure so we can collaborate on
maintaining it.

This is a first step in that direction, having more up-to-date
reports hosted from GitHub Pages. Next steps might be
to generate those reports from nix-community infra and
finding a good way to also host the (possibly huge)
diffoscope reports, but let's take one step at a time.