terraform/fastlylog: add my own S3 bucket
I would like to request access to the S3 logs as I would like to conduct
work to expose open data to the NixOS community regarding the access pattern
to our Fastly CDN, this is performed in the context of bringing our own self hosted
binary cache.

But also as part of more long term work to offer self-service data to our community
so they can work on popcount, and more.

I am not expert enough in Terraform to write stuff that works on the first try
without being able to `terraform init`, so I'd love to get help from admin folks here.

I am also open to reduce the scope of the data I receive, I was planning to postprocess
them anyway to do proper aggregation to keep interesting data but not expose any PII data,
obviously.

If that's not okay, let me know what we can do to aim for this purpose.