hetzner-01: rename to caliban.nixos.org
We needed a name to create CNAMES to eventually, so here is one
proposal, since we're going for pet names.

Caliban is the second largest moon of Uranus, and should somewhat fit
into the existing naming scheme.

https://en.wikipedia.org/wiki/Caliban_(moon)

Once we have the DNS name deployed we can update the deployment address
accordingly.