POC: add peribolos configuration
This is a POC to see how the
[peribolos](https://github.com/uwu-tools/peribolos) config could work for us.

The idea would be that anybody can submit a PR to change the NixOS org
configuration, making the changes more transparent and available.

The peribolos/config.yaml contains a dump of the NixOS and NixCon orgs
and needs to be cleaned up a bit. Notably, the list of org members and
nixpkgs committers needs to be removed as it is handled by a separate
tool.

My first impression so far:
* The config file is too large. It would be good if it could be split
  up.
* It doesn't handle Projects (requested by Raito).

Otherwise it looks pretty good so far.

I haven't tested the application of changes yet, this needs to be tested
on another org first.

The question of who or what would be applying the changes is also up in
the air (given that the token has so much power).