haumea: restore backups
Backups to rsync.net stopped working on Jan 10th, when haumea stopped
being able to connect to the sync. Then on Jan 13th it was able to
connect to the sync again, but incompatibilities in the zrepl versions
on both sides made it impossible to send backups.

We unfortunately don't control the rsync.net account, so for now drop it
in favor of a backup on one of my personal hosts. Not meant to be a long
term solution (famous last words).