non-critical-infra: init jitsi
This commit sets up jitsi for the nixos-org.
The jitsi config has the lobby enabled for every-room by-default,
this means only people that get access will be able to join.
Access control is done using the `Secure Domain Lobby Bypass` plugin and
should be changed to `Token Lobby Bypass` once we have a IDP.