Use `diskSelector` variable
This doesn't make a material difference (the only value for
`diskSelector` is `/`, which we were already using here), but somebody
went to the effort to put this in a variable, so IMO we should be
consistent about using it.

An alternative would be to just inline the mountpoint everywhere. Let me
know what you prefer.