Use /experimental-installer instead of /nix
Redirect from /experimental-installer to the experimental installer
releases instead of /nix. The installer is still experimental so that
should be clear from the URL.

Also add a few redirects to make URLs nicer.

Redirect /experimental-nix to the latest GitHub release script,
/experimental-nix/* to the latest GitHub arch specific installers, and
use /experimental-nix/tag/$version/*
to redirect to specific releases.

This allows for a shorter URL for the most common install instruction,
and it more closely mirrors the URL structure used by DetSys which is
upstream for experimental-nix-installers

https://artifacts.nixos.org/experimental-installer -> https://github.com/NixOS/experimental-nix-installer/releases/latest/download/nix-installer.sh
https://artifacts.nixos.org/experimental-installer/* -> https://github.com/NixOS/experimental-nix-installer/releases/latest/download/*
https://artifacts.nixos.org/experimental-installer/tag/* -> https://github.com/NixOS/experimental-nix-installer/releases/download/*