fix: Reflected XSS on channels.nixos.org and releases.nixos.org
Additional hardening such as adding a CSP header will be done
later on but it requires some additional upstream to be useful.

Upstream already merged the fix we submitted:
https://github.com/rufuspollock/s3-bucket-listing/commit/57ed3fa7be308ac82cfa18c99e207f1d1dc9ef74

Upgrading to the latest version of the s3-bucket-listing script seems to
require additional changes and the security fix is small. It seems
preferable to only deal with the security fix at this point and deal
with the upgrade and potential complications later on.

Fixes GHSA-3xg5-p8ch-832g